+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.0.44 (March 20, 2018)
 - Fixed persistent Jade mismatch.
+- Highlighted doc comments in TypeScript.
 
 ## 0.0.43 (March 20, 2018)
 - More work on templating languages. (Jade, Twig, EJS).