+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # ChangeLog
 
+## 0.0.4% (March 20, 2018)
+- Treat single & double quoted object keys as object keys in JavaScript.
+
 ## 0.0.44 (March 20, 2018)
 - Fixed persistent Jade mismatch.
 - Highlighted doc comments in TypeScript.